bundler: Extract pydantic model definitions from the parser module
This is a cosmetic change. Model definitions should be mostly static
and should be parser agnostic and so move the model definitions out of
the parser logic to a new module.
